# backend/docs/material/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the material tracking application's api for dealing with the materials",
        "title": "Material Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Comms group",
            "url": "http://github.com/PatheticApathy/CoMMS"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/checkouts/all": {
            "get": {
                "description": "gets all checkout logs if they exist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkout logs"
                ],
                "summary": "fetches all checkout logs",
                "responses": {
                    "200": {
                        "description": "checkout logs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/materialdb.CheckoutLog"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/checkouts/in": {
            "put": {
                "description": "Adds material log  to the database using valid json structure",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkout logs"
                ],
                "summary": "post material log to database",
                "parameters": [
                    {
                        "description": "id of checkoutlog",
                        "name": "logid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "checkout log",
                        "schema": {
                            "$ref": "#/definitions/materialdb.CheckoutLog"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/checkouts/out": {
            "post": {
                "description": "Adds checkout log for a materials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checkout logs"
                ],
                "summary": "post checkout of material",
                "parameters": [
                    {
                        "description": "Format of add ckeckout log",
                        "name": "checkoutlg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/materialdb.AddCheckoutLogParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "checkout log",
                        "schema": {
                            "$ref": "#/definitions/materialdb.CheckoutLog"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/sites/add": {
            "post": {
                "description": "Adds job_site to the database using valid json structure",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "post job_site to database",
                "parameters": [
                    {
                        "description": "Format of add jobsite request",
                        "name": "jobsite",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/materialdb.AddJobSiteParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "jobsite",
                        "schema": {
                            "$ref": "#/definitions/materialdb.JobSite"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sites/all": {
            "get": {
                "description": "Get all jobsites",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "fetches all job_sites",
                "responses": {
                    "200": {
                        "description": "job site",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/materialdb.JobSite"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sites/search": {
            "get": {
                "description": "Gets jobsites using id(may add more parameters later)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "fetches job_site based on given paremeters",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "jobsite's identification number",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "job site",
                        "schema": {
                            "$ref": "#/definitions/materialdb.JobSite"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/add": {
            "post": {
                "description": "The use defined in the material database is only the bear essentials for logging",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "adds user to the material database",
                "parameters": [
                    {
                        "description": "Format of add user request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/materialdb.AddUserParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user",
                        "schema": {
                            "$ref": "#/definitions/materialdb.User"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/all": {
            "get": {
                "description": "gets all users from db",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "gets all users from the material database",
                "responses": {
                    "200": {
                        "description": "user",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/materialdb.User"
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "description": "gets user from db",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "get user from the material database",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user",
                        "schema": {
                            "$ref": "#/definitions/materialdb.User"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "materialdb.AddCheckoutLogParams": {
            "type": "object",
            "properties": {
                "item_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "materialdb.AddJobSiteParams": {
            "type": "object",
            "properties": {
                "addr": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "location_lat": {
                    "$ref": "#/definitions/sql.NullFloat64"
                },
                "location_lng": {
                    "$ref": "#/definitions/sql.NullFloat64"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "materialdb.AddUserParams": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "site_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                }
            }
        },
...
        "materialdb.CheckoutLog": {
            "type": "object",
            "properties": {
                "checkin_time": {
                    "type": "string"
                },
                "checkout_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "materialdb.JobSite": {
            "type": "object",
            "properties": {
                "addr": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "id": {
                    "type": "integer"
                },
                "location_lat": {
                    "$ref": "#/definitions/sql.NullFloat64"
                },
                "location_lng": {
                    "$ref": "#/definitions/sql.NullFloat64"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "materialdb.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "site_id": {
                    "$ref": "#/definitions/sql.NullInt64"
                }
            }
        },
        "sql.NullFloat64": {
            "type": "object",
            "properties": {
                "float64": {
                    "type": "number"
                },
                "valid": {
                    "description": "Valid is true if Float64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
...
}