# backend/pkg/databases/userdb/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package userdb

type User struct {
	ID        int64
	Username  string
	Password  string
	Firstname string
	Lastname  string
	Company   string
	Site      string
	Role      string
	Email     string
	Phone     string
}